+++ src/compiler/parse_global.c
@@ -48,8 +48,9 @@ void recover_top_level(ParseContext *c)
 			case TOKEN_EXTERN:
 			case TOKEN_ENUM:
 			case TOKEN_ALIAS:
-			case TOKEN_DEF:
-			case TOKEN_FAULT:
+			case TOKEN_TYPEDEF:
+			case TOKEN_ATTRDEF:
+			case TOKEN_FAULTDEF:
 				return;
 			case TOKEN_CONST:
 			case TOKEN_ASM:
@@ -1679,11 +1680,11 @@ static bool parse_struct_body(ParseContext *c, Decl *parent)
 
 
 /**
- * distinct_declaration ::= 'distinct' TYPE_IDENT opt_interfaces '=' 'inline'? type ';'
+ * typedef_declaration ::= 'typedef' TYPE_IDENT opt_interfaces attributes? '=' 'inline'? type ';'
  */
-static inline Decl *parse_distinct_declaration(ParseContext *c)
+static inline Decl *parse_typedef_declaration(ParseContext *c)
 {
-	advance_and_verify(c, TOKEN_DISTINCT);
+	advance_and_verify(c, TOKEN_TYPEDEF);
 
 	Decl *decl = decl_new_with_type(symstr(c), c->span, DECL_DISTINCT);
 
@@ -1701,7 +1702,7 @@ static inline Decl *parse_distinct_declaration(ParseContext *c)
 
 	if (tok_is(c, TOKEN_FN))
 	{
-		PRINT_ERROR_HERE("A distinct type cannot define a new function type, but you can make a distinct type from an existing function type, e.g. `def FooFn = fn void(); distinct Bar = FooFn;`");
+		PRINT_ERROR_HERE("A distinct type cannot define a new function type, but you can make a distinct type from an existing function type, e.g. `alias FooFn = fn void(); typedef Bar = FooFn;`");
 		return poisoned_decl;
 	}
 	// 2. Now parse the type which we know is here.
@@ -1928,7 +1929,7 @@ static inline void decl_add_type(Decl *decl, TypeKind kind)
 	decl->type = type;
 }
 /**
- * typedef_declaration ::= ALIAS TYPE_IDENT '=' typedef_type ';'
+ * typedef_declaration ::= ALIAS TYPE_IDENT attributes? '=' typedef_type ';'
  *
  * typedef_type ::= func_typedef | type generic_params?
  * func_typedef ::= 'fn' optional_type parameter_type_list
@@ -1958,6 +1959,8 @@ static inline Decl *parse_alias_type(ParseContext *c)
 		return poisoned_decl;
 	}
 
+	if (!parse_attributes_for_global(c, decl)) return poisoned_decl;
+
 	CONSUME_OR_RET(TOKEN_EQ, poisoned_decl);
 
 	// 1. Did we have `fn`? In that case it's a function pointer.
@@ -1974,7 +1977,6 @@ static inline Decl *parse_alias_type(ParseContext *c)
 		{
 			return poisoned_decl;
 		}
-		if (!parse_attributes_for_global(c, decl)) return poisoned_decl;
 		RANGE_EXTEND_PREV(decl_type);
 		RANGE_EXTEND_PREV(decl);
 		CONSUME_EOS_OR_RET(poisoned_decl);
@@ -2015,7 +2017,6 @@ static inline Decl *parse_alias_type(ParseContext *c)
 	{
 		decl->typedef_decl.is_redef = true;
 	}
-	if (!parse_attributes_for_global(c, decl)) return poisoned_decl;
 
 	RANGE_EXTEND_PREV(decl);
 	CONSUME_EOS_OR_RET(poisoned_decl);
@@ -2023,7 +2024,7 @@ static inline Decl *parse_alias_type(ParseContext *c)
 }
 
 /**
- * define_ident ::= 'define' (IDENT | CONST_IDENT | AT_IDENT) '=' identifier_alias generic_params?
+ * define_ident ::= 'define' (IDENT | CONST_IDENT | AT_IDENT) attributes? '=' identifier_alias generic_params?
  *
  * identifier_alias ::= path? (IDENT | CONST_IDENT | AT_IDENT)
  */
@@ -2064,6 +2065,9 @@ static inline Decl *parse_alias_ident(ParseContext *c)
 	}
 	// 4. Advance and consume the '='
 	advance(c);
+
+	if (!parse_attributes_for_global(c, decl)) return poisoned_decl;
+
 	CONSUME_OR_RET(TOKEN_EQ, poisoned_decl);
 
 	if (tok_is(c, TOKEN_FN))
@@ -2073,7 +2077,6 @@ static inline Decl *parse_alias_ident(ParseContext *c)
 
 	ASSIGN_EXPR_OR_RET(decl->define_decl.alias_expr, parse_expr(c), poisoned_decl);
 
-	if (!parse_attributes_for_global(c, decl)) return poisoned_decl;
 
 	RANGE_EXTEND_PREV(decl);
 	CONSUME_EOS_OR_RET(poisoned_decl);
@@ -2124,7 +2127,7 @@ static inline Decl *parse_alias_attribute(ParseContext *c)
 }
 
 /**
- * define_decl ::= ALIAS define_type_body |
+ * define_decl ::= ALIAS define_type_body
  */
 static inline Decl *parse_alias(ParseContext *c)
 {
@@ -2255,11 +2258,11 @@ static inline Decl *parse_fault(ParseContext *c)
 }
 
 /**
- * fault_declaration ::= FAULT CONST_IDENT (',' CONST_IDENT)* ','? ';'
+ * faultdef_declaration ::= FAULTDEF CONST_IDENT (',' CONST_IDENT)* ','? ';'
  */
-static inline Decl *parse_fault_declaration(ParseContext *c)
+static inline Decl *parse_faultdef_declaration(ParseContext *c)
 {
-	advance_and_verify(c, TOKEN_FAULT);
+	advance_and_verify(c, TOKEN_FAULTDEF);
 
 	if (c->lexer.token_type == TOKEN_EOS)
 	{
@@ -2998,9 +3001,9 @@ Decl *parse_top_level_statement(ParseContext *c, ParseContext **context_out)
 			if (has_real_contracts) goto CONTRACT_NOT_ALLOWED;
 			decl = parse_interface_declaration(c);
 			break;
-		case TOKEN_DISTINCT:
+		case TOKEN_TYPEDEF:
 			if (has_real_contracts) goto CONTRACT_NOT_ALLOWED;
-			decl = parse_distinct_declaration(c);
+			decl = parse_typedef_declaration(c);
 			break;
 		case TOKEN_CONST:
 			if (has_real_contracts) goto CONTRACT_NOT_ALLOWED;
@@ -3018,9 +3021,9 @@ Decl *parse_top_level_statement(ParseContext *c, ParseContext **context_out)
 			if (has_real_contracts) goto CONTRACT_NOT_ALLOWED;
 			decl = parse_enum_declaration(c);
 			break;
-		case TOKEN_FAULT:
+		case TOKEN_FAULTDEF:
 			if (has_real_contracts) goto CONTRACT_NOT_ALLOWED;
-			decl = parse_fault_declaration(c);
+			decl = parse_faultdef_declaration(c);
 			break;
 		case TOKEN_IDENT:
 			if (has_real_contracts) goto CONTRACT_NOT_ALLOWED;